refactor(admin-config): Move asset resolution logic to `AdminConfig`
Move the logic to ensure that asset directories end with a trailing
slash, to resolve global asset references, and to add missing plugin
definitions to the `AdminConfig` class. This ensures that report
definitions are always correctly configured, also if the `AdminConfig`
is not created via the `AdminConfigResolver`.
