+++ components/admin-config/backend/src/main/kotlin/AdminConfig.kt
@@ -54,7 +54,7 @@ class AdminConfig(
          * reports. Typically, some reports should be configured.
          */
         val DEFAULT_REPORTER_CONFIG = ReporterConfig(
-            reportDefinitionsMap = ReporterConfig.addDefinitionsForUnreferencedPlugins(emptyMap()),
+            reportDefinitionsMap = emptyMap(),
             howToFixTextProviderFile = ORT_HOW_TO_FIX_TEXT_PROVIDER_FILENAME,
             customLicenseTextDir = null
         )
